Remove test_model.py file that was re-created during rebasing
