+++ tests/test_book_model.py
@@ -1,4 +1,5 @@
 from app.models.book import Book
+import pytest
 
 def test_to_dict_no_missing_data():
     # Arrange
@@ -56,3 +57,55 @@ def test_to_dict_missing_description():
     assert result["id"] == 1
     assert result["title"] == "Ocean Book"
     assert result["description"] is None
+
+def test_from_dict_returns_book():
+    # Arrange
+    book_data = {
+        "title": "New Book",
+        "description": "The Best!"
+    }
+
+    # Act
+    new_book = Book.from_dict(book_data)
+
+    # Assert
+    assert new_book.title == "New Book"
+    assert new_book.description == "The Best!"
+
+def test_from_dict_with_no_title():
+    # Arrange
+    book_data = {
+        "description": "The Best!"
+    }
+
+    # Act & Assert
+    with pytest.raises(KeyError, match = 'title'):
+        new_book = Book.from_dict(book_data)
+
+def test_from_dict_with_no_description():
+    # Arrange
+    book_data = {
+        "title": "New Book"
+    }
+
+    # Act & Assert
+    with pytest.raises(KeyError, match = 'description'):
+        new_book = Book.from_dict(book_data)
+
+def test_from_dict_with_extra_keys():
+    # Arrange
+    book_data = {
+        "extra": "some stuff",
+        "title": "New Book",
+        "description": "The Best!",
+        "another": "last value"
+    }
+    
+    # Act
+    new_book = Book.from_dict(book_data)
+
+    # Assert
+    assert new_book.title == "New Book"
+    assert new_book.description == "The Best!"
+    assert hasattr(new_book, "extra") == False
+    assert hasattr(new_book, "another") == False
